requirements.txt
flask
pyotp
qrcode
fpdf
pillow
